Delete commented-out callback and remove waiting for timeout. That's not necessary because we just insert the transmit request into the FIFO and off we go.
Former-commit-id: 754fd4aa642bb73a29cfa06e2227e15a0331e854

+++ libraries/CAN/src/CAN.h
@@ -45,7 +45,7 @@ public:
   bool begin(CanMtuSize const can_mtu_size);
   void end();
 
-  uint8_t write(CanMsg const & msg);
+  int write(CanMsg const & msg);
   uint8_t read(CanMsg & msg);
 
   bool tx_complete;
